MDEV-17262: mysql crashed on galera while node rejoined cluster
This patch contains a fix for the MDEV-17262/17243 issues and
new mtr test.

These issues (MDEV-17262/17243) have two reasons:

1) A transaction loses its status as "transaction registered
in the MySQL for 2pc coordinator" (in InnoDB) after an intermediate
commit due to the fact that since version 10.2 the write_row() function
(from the ha_innodb.cc) does not call trx_register_for_2pc(m_prebuilt->trx);
during the processing of split transactions. It is necessary to restore
this call inside the write_row() when an intermediate commit was made
(for a split transaction).

Similarly, we need to restore the flag of the started transaction
(m_prebuilt-> sql_stat_start); after intermediate commit.

Calling the table->file->extra(HA_EXTRA_FAKE_START_STMT) inside
the wsrep_load_data_split() function (in sql_load.cc), will also
do this, but it will be too late, as a result, the call to the
wsrep_append_keys() function inside the InnoDB engine may be lost
or function may be called with invalid transaction identifier.

2) If a transaction with the LOAD DATA statement is divided into
logical mini-transactions (of the 10K rows) and binlog is rotated, then
in rare cases due to the wsrep handler re-registration at the boundary
of the split, the last portion of data may be lost. Since splitting
of the LOAD DATA into mini-transactions is logical, I believe that
we should not allow these mini-transactions to fall into separate
binlogs. Therefore, it is necessary to prohibit the rotation of
binlog in the middle of processing LOAD DATA.

https://jira.mariadb.org/browse/MDEV-17262 and
https://jira.mariadb.org/browse/MDEV-17243