MDEV-17835: Remove wsrep-sst-method=xtrabackup
The second line of changes related to replacing xtrabackup with
mariabackup: removed all unnecessary references to xtrabackup in
the documentation, in some comments, in the files that are used to
prepare the packages. Corrected tests from the galera_3nodes suite
that mentioned xtrabackup or the old (associated with xtrabackup)
version of innobackupex. Fixed flaws in galera_3nodes mtr suite
tests and control scripts, because of which they could not work
with mariabackup.

https://jira.mariadb.org/browse/MDEV-17835